pyproject.toml
[tool.poetry]
name = "mo-tech"
version = "0.1.0"
description = ""
authors = ["Seba Zapata <s.zapata05@outlook.com>"]
readme = "README.md"
packages = [{include = "mo_tech"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
djangorestframework = "^3.14.0"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
